+++ modules/tinymce/src/plugins/fullpage/main/ts/Plugin.ts
@@ -14,8 +14,8 @@ import * as FilterContent from './core/FilterContent';
 import * as Buttons from './ui/Buttons';
 
 export default () => {
-  console.warn('Notice: The fullpage plugin has been marked for removal from TinyMCE 6.0');
   PluginManager.add('fullpage', (editor) => {
+    console.warn('The fullpage plugin has been deprecated and marked for removal in TinyMCE 6.0');
     const headState = Cell(''), footState = Cell('');
 
     Commands.register(editor, headState);